frontend/requirements.txt
streamlit
python-dotenv
requests
supabase
